+++ shared/components/ui/mantine/elements/buttons/unstyledButton/v1.0.0/types.d.ts
@@ -1,0 +1,7 @@
+import { BaseReactProps } from '@shared/node'
+import { Scope } from '@shared/scope'
+
+export type Props = BaseReactProps & {
+    useScope: boolean
+    scope?: Scope    
+}