+++ shared/components/ui/mantine/molecules/flex/v1.0.0/types.d.ts
@@ -1,0 +1,11 @@
+import { FlexProps } from '@mantine/core'
+import { BaseReactProps } from '@shared/node'
+
+export type Props = {
+    flexAlign?: FlexProps['align']
+    flexJustify?: FlexProps['justify']
+    flexDirection?: FlexProps['direction']
+    flexWrap?: FlexProps['wrap']
+    customProps?: any
+    children?: any
+}