+++ shared/components/ui/mantine/molecules/paper/v1.0.0/paper.tsx
@@ -1,0 +1,14 @@
+import { forwardRef } from "react"
+import { Props } from "./types";
+import React from "react";
+import { Paper } from "@mantine/core";
+
+export default forwardRef(function (props: Props) {
+
+    return <Paper
+        {...props}
+        {...props.customProps}
+    >
+        {props.children}
+    </Paper>
+})