+++ shared/internal/port/v0.6.0/src/ports/style.ts
@@ -8,4 +8,6 @@ export default [
     { name: 'backgroundColor', displayName: 'Bg color', group: 'Style', type: 'string' },
     { name: 'color', displayName: 'Color', group: 'Style', type: 'string' },
     { name: 'opacity', displayName: 'Opacity', group: 'Style', type: 'number', default: 1 },
+    { name: 'shadow', displayName: 'Shadow', group: 'Style', type: getEnumType(enums.sizes), default: 'sm' },
+    { name: 'withBorder', displayName: 'With border', group: 'Style', type: 'boolean', },
 ] as const satisfies readonly NodePort[]